test_package/requirements.txt
airtest==1.2.7
pocoui==1.0.88
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.0.12
decorator==5.1.1
Deprecated==1.2.13
facebook-wda==1.4.6
hrpc==1.0.9
idna==3.4
Jinja2==3.0.3
MarkupSafe==2.0.1
mss==6.1.0
numpy==1.19.5
opencv-contrib-python==3.4.2.17
Pillow==8.4.0
py==1.11.0
python-xlib==0.32
pywinauto==0.6.3
requests==2.27.1
retry==0.9.2
six==1.16.0
urllib3==1.26.12
websocket-client==0.48.0
wrapt==1.14.1
